# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Observatorio ULEAM API",
        "description": "API para el Sistema de Gestión y Visualización de Datos Universitarios",
        "contact": {
            "name": "ULEAM",
            "email": "soporte@uleam.edu.ec"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/departamentos": {
            "get": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Listar departamentos",
                "description": "Obtiene la lista de departamentos del usuario autenticado",
                "operationId": "8d5827a5011b0ae66c95a7298b27f775",
                "responses": {
                    "200": {
                        "description": "Lista de departamentos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "departamentos": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Crear departamento",
                "description": "Crea un nuevo departamento",
                "operationId": "87968a98b4b69584882201d7fefb7423",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Bienestar Estudiantil"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Departamento de bienestar"
                                    },
                                    "publico": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Departamento creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/departamentos/{id}": {
            "get": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Obtener departamento",
                "description": "Obtiene un departamento específico",
                "operationId": "72b7b539338fdb93139fe4aedcc8cb5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Departamento encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Actualizar departamento",
                "description": "Actualiza un departamento existente",
                "operationId": "7853ccb9c6dea5c231e9f6c74b06dd0c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "publico": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Departamento actualizado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Departamentos"
                ],
                "summary": "Eliminar departamento",
                "description": "Elimina un departamento (soft delete)",
                "operationId": "b7a36706003080e62a0e4f709aa3ed17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Departamento eliminado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Ingresa el token en formato: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}